% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmaclip.R
\name{lemmaclip}
\alias{lemmaclip}
\title{Filter lemmas to clipboard}
\usage{
lemmaclip(lower, upper = 50000)
}
\arguments{
\item{lower}{positive int. lowest NT vocab freq. to include.}

\item{upper}{positive int. highest NT vocab freq. to include.}
}
\value{
clips information to clipboard. no important return.
}
\description{
Filter lemmas to clipboard
}
\examples{
lemmaclip(25)
}
