% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmaclip.R
\name{lemmaclip}
\alias{lemmaclip}
\title{Filter lemmas to clipboard}
\usage{
lemmaclip(lower, upper = 50000)
}
\arguments{
\item{lower}{positive int. lowest NT vocab freq. to include.}

\item{upper}{positive int. include NT vocab BELOW this frequency (i.e. not this frequency).}
}
\value{
clips lemma query to clipboard. Returns tibble of clipped lemmas with relevant data
}
\description{
Filter lemmas to clipboard
}
\examples{
lemmaclip(25)
}
